Refactor selection rectangles into dedicated pass
